Sending "big" files is back again :)
PostMessageRequest pack is prepared in-memory with the entire content of
the file being sent. This is only a temporary fix for this and actually
the protobuf library should be changed to behave in a more "streamed"
fashion.
Next change will be limiting the size of a message to a limit imposed
by server configuration (via serverinfo).

Signed-off-by: Daniele Ricci <daniele.athome@gmail.com>