Import key pair process (preference not active yet)
This is a very basic and ugly import key pair process which uses the same
procedure of key regeneration, but instead it sends the imported key to the
server, which will accept it as valid.
Actual user interaction is still missing, it will be implemented soon.
After that, importing will be implemented also for the registration phase.

Signed-off-by: Daniele Ricci <daniele.athome@gmail.com>