Make adaptive ping actually works and decrease min interval to 90 seconds
Experiments on 3G/4G networks showed that below 90 seconds
connection remains stable. Just passing the 90 seconds threshold,
the connection is quietly destroyed by the carrier.

There are still some glitches in this adaptive algorithm (namely:
learning stickness, that is, the algorithm tends to forget soon
what was the best value, I'll have to work on that).

Signed-off-by: Daniele Ricci <daniele.athome@gmail.com>