Fix adaptive ping issue
Fixed an issue that was causing the adaptive ping algorithm to not consider
the backoff interval for trying to increase the ping interval.
Now the adaptive ping manager keeps track of the ping streak (i.e. how much
time we have been having successful pings) and uses it to determine how and
when try an increase of the ping interval.

Signed-off-by: Daniele Ricci <daniele.athome@gmail.com>