Contextual reply (database and protocol parts; #594)
The referenced message is not stored in our database, but is rather
referenced through the message id received in the forwarded stanza.
This is to save space and avoid storing too much stuff in our database,
but also to keep consistency. This has the side effect of not being
able to show the original message in a reply if it is deleted.

Signed-off-by: Daniele Ricci <daniele@casaricci.it>