Be sure to keep message center alive while connecting or reconnecting
Keeping the wakelock is not enough because we need to block the internal
idle management too. In this case, we increment the reference counter one
time before creating the connection and then we decrement it after
authenticating or when disconnecting. This way, even under stress conditions
(e.g. bad network quality), the message center will never shutdown because
of inactivity.

Signed-off-by: Daniele Ricci <daniele@casaricci.it>