Introducing the message center client
In an attempt to simplify interaction between UI and the message
center, a new class is now entering the field: MessageCenterClient.
It will provide (and in part already does) a transparent listener
interface to use the message center, avoiding cluttering and ugly
broadcast receiver stuff in the UI.
This will have of course a cost, since several data structures will
appear, but it will be a minor overhead - negligible probably.
IMO I believe this will bring benefits for the code itself, and it
will make the code more testable.

Signed-off-by: Daniele Ricci <daniele@casaricci.it>