Fix possible race condition between message receive and inactivity shutdown
It may happen that, during MessageListener processing, the message center decides
it's time to shut down for inactivity. During that time, mConnection object in the
message center could become null at any time, driving the MessageListener into
oblivion as a NullPointerException is generated (not crashing, but logged by Smack
thus invisible to our internal file logger - I need to fix that by the way).

By storing the connection object at creation time (in a weak reference), we can be
sure the object is never null and message processing can be always fulfilled.

Signed-off-by: Daniele Ricci <daniele@casaricci.it>