Hack: do not convert localpart to lowercase
In order to maintain compatibility with current clients, I made a custom
stringprep class which skips the toLowerCase call for the localpart.
This is needed for Kontalk group JIDs where the group ID (which becomes
the localpart) may have capital letters (it's randomly generated).

Until a better solution is implemented, but it's quite the work.

Signed-off-by: Daniele Ricci <daniele@casaricci.it>