Disable write external storage permission on API level 29 (#1291)
We are now ready to not ask for write external storage permission on API
level 29 (Android R). With scoped storage, MediaStore access will be
public so no need for a permission.
Read external storage permission will stay for the import messages
database function (that's still a debug feature anyway for now).

Signed-off-by: Daniele Ricci <daniele@casaricci.it>